exclude rejected changesets from review reports
If someone has voted the minimum on a non-optional label, the change
is considered "rejected" and does not need any further review work
until the rejection is removed or a new patch set is submitted. This
includes, for Storj purposes, the case where the build bot has added
a Verified=-1 vote. So, in short, changes that failed the built won't
show up as needing reviews anymore.